requirements.txt
python==3.6
tensorly==0.4.5
pytest==6.2.2
shapely==1.7.1
scikit-learn==0.24.1
scikit-image==0.17.2
numpy==1.19.5
tqdm==4.59.0
geopandas
netcdf4==1.5.6
oct2py==5.2.0
pyyaml==5.4.1
matplotlib==3.3.4
loguru==0.5.3
ray==1.3.0
pandas==1.1.5
